# mem: implement `ReadAll()` for more efficient `io.Reader` consumption