I moved my project to gRPC 1.66.2 and saw a good reduction in RAM consumption. Now the hot spot is `decompress()`, where `io.Copy()` allocates a temporary buffer, reads from the reader into it, copies the read data into another buffer it got from the pool. This is an unnecessary allocation, an unnecessary copy, and underutilized buffers from the pool.

This PR adds `mem.ReadAll()` (like `io.ReadAll()`) to efficiently consume a reader into buffers from the pool.

<img width="331" alt="Screenshot 2024-09-19 at 8 11 15 AM" src="https://github.com/user-attachments/assets/a38c0ba7-a3b4-4f89-90a6-b0ec367fddae">

I found https://github.com/grpc/grpc-go/issues/7631 while working on this code (I have similar code in my project, but decided to contribute it upstream and replace this `io.Copy` with it).

RELEASE NOTES:
- mem: implement a `ReadAll()` method for more efficient `io.Reader` consumption