@ash2k 2024-10-29
Related question: I see quite a few calls with a `nil` `pool` - `mem.NewBuffer(&someDataSlice, nil)`. Why not swap this with a type cast like this `mem.SliceBuffer(someDataSlice)`? No need to call a function.