@easwars 2024-10-30
> Related question: I see quite a few calls with a `nil` `pool` - `mem.NewBuffer(&someDataSlice, nil)`. Why not swap this with a type cast like this `mem.SliceBuffer(someDataSlice)`? No need to call a function.

IIRC, the `SliceBuffer` type was added a little later on during the review process on the PR where all the buffering functionality was added, and there is a good chance some callsites were not fixed. I don't see a reason why we should be opposed to doing that change and would happy review a PR with that change. Thanks.