@dfawley 2024-11-12
Thanks for the PR!  That's a nice performance gain.